requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
Scrapy==0.24.4
scrapyd==1.0.1
scrapy-heroku==0.7.1

akismet==0.2.0
amqp==1.4.9
anyjson==0.3.3
attrs==16.0.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
boto==2.39.0



bs4==0.0.1
celery==3.1.23
certifi==14.5.14
cffi==1.7.0
chardet==2.3.0
cryptography==1.4
cssselect==0.9.2
decorator==4.0.10
defusedxml==0.4.1

django-celery==3.1.17
django-comments-xtd==1.5.3
django-contrib-comments==1.7.1
django-crispy-forms==1.6.0
django-disqus==0.5
django-dynamic-scraper==0.11.0
django-fluent-comments==1.2.1
django-localflavor==1.0
django-markdown-deux==1.0.5
django-modalview==0.1.5
django-pagedown==0.1.1
django-registration==2.1
django-registration-redux==1.3
django-secure==1.0.1
django-sslify==0.2
django-sslserver==0.19
django-storages==1.1.8
django-tag-parser==2.1

python-postmark==0.3.1

docutils==0.12
enum34==1.1.6
funcsigs==1.0.2
future==0.15.2

idna==2.1
ipaddress==1.0.16
jsonpath-rw==1.4.0
kombu==3.0.35
lxml==3.4.0
markdown2==2.3.1
mock==2.0.0
numpy==1.8.2
parsel==1.0.2
pbr==1.10.0
Pillow==3.1.0
ply==3.8
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyOpenSSL==16.0.0


pytz==2015.7
queuelib==1.4.2
scrapy-djangoitem==1.1.1
service-identity==16.0.0
six==1.10.0
South==1.0.2
stevedore==1.15.0
Twisted==16.2.0


w3lib==1.14.2
